removes obsolete tests and moves the ones we want to keep over to other test files
